pyproject.toml
[project]
name = "ticktick-to-sqlite"
version = "0.1"
description = "Import TickTick data into a SQLite database"
readme = "README.md"
authors = [{name = "Scott Carvalho"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "ticktick-py",
    "sqlite-utils",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/Scarvy/ticktick-to-sqlite"
Changelog = "https://github.com/Scarvy/ticktick-to-sqlite/releases"
Issues = "https://github.com/Scarvy/ticktick-to-sqlite/issues"
CI = "https://github.com/Scarvy/ticktick-to-sqlite/actions"

[project.entry-points.console_scripts]
ticktick-to-sqlite = "ticktick_to_sqlite.cli:cli"

[project.optional-dependencies]
test = [
    "pytest", 
    "ruff"
]
